move mesh context builder to context
