+++ pkg/xds/cache/mesh/cache_test.go
@@ -99,7 +99,7 @@ var _ = Describe("MeshSnapshot Cache", func() {
 		lookupIPFunc := func(s string) ([]net.IP, error) {
 			return []net.IP{net.ParseIP(s)}, nil
 		}
-		meshContextBuilder := mesh.NewMeshContextBuilder(lookupIPFunc, "zone-1", vips.NewPersistence(core_manager.NewResourceManager(s), manager.NewConfigManager(s)), "mesh")
+		meshContextBuilder := xds_context.NewMeshContextBuilder(lookupIPFunc, "zone-1", vips.NewPersistence(core_manager.NewResourceManager(s), manager.NewConfigManager(s)), "mesh")
 		meshCache, err = mesh.NewCache(
 			countingManager,
 			expiration,
