+++ pkg/xds/topology/dataplanes_test.go
@@ -1,8 +1,9 @@
-package topology
+package topology_test
 
 import (
 	"net"
 
+	"github.com/kumahq/kuma/pkg/xds/topology"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/pkg/errors"
@@ -43,7 +44,7 @@ var _ = Describe("Resolve Dataplane address", func() {
 			}
 
 			// when
-			resolvedDp, err := ResolveAddress(lif, dp)
+			resolvedDp, err := topology.ResolveAddress(lif, dp)
 
 			// then
 			Expect(err).ToNot(HaveOccurred())
@@ -70,7 +71,7 @@ var _ = Describe("Resolve Dataplane address", func() {
 				{Spec: &mesh_proto.Dataplane{Networking: &mesh_proto.Dataplane_Networking{Address: "192.168.1.2", AdvertisedAddress: "192.0.2.2"}}},
 			}
 
-			actual := ResolveAddresses(core.Log, lif, given)
+			actual := topology.ResolveAddresses(core.Log, lif, given)
 			Expect(actual).To(HaveLen(4))
 			Expect(actual).To(Equal(expected))
 		})
@@ -88,7 +89,7 @@ var _ = Describe("Resolve Dataplane address", func() {
 				{Spec: &mesh_proto.Dataplane{Networking: &mesh_proto.Dataplane_Networking{Address: "192.168.1.1"}}},
 				{Spec: &mesh_proto.Dataplane{Networking: &mesh_proto.Dataplane_Networking{Address: "192.168.1.2", AdvertisedAddress: "192.0.2.2"}}},
 			}
-			actual := ResolveAddresses(core.Log, lif, given)
+			actual := topology.ResolveAddresses(core.Log, lif, given)
 			Expect(actual).To(HaveLen(3))
 			Expect(actual).To(Equal(expected))
 		})
