+++ pkg/xds/topology/timeouts_test.go
@@ -1,9 +1,10 @@
-package topology
+package topology_test
 
 import (
 	"context"
 	"time"
 
+	"github.com/kumahq/kuma/pkg/xds/topology"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
@@ -63,7 +64,7 @@ var _ = Describe("Timeout", func() {
 			err := rm.Create(ctx, timeoutsFrontendV1ToRedis, store.CreateBy(model.MetaToResourceKey(timeoutsFrontendV1ToRedis.GetMeta())))
 			Expect(err).ToNot(HaveOccurred())
 
-			timeoutMap, err := GetTimeouts(ctx, dataplane, rm)
+			timeoutMap, err := topology.GetTimeouts(ctx, dataplane, rm)
 			Expect(err).ToNot(HaveOccurred())
 			Expect(timeoutMap).To(HaveLen(1))
 			Expect(timeoutMap).To(HaveKeyWithValue(
@@ -85,7 +86,7 @@ var _ = Describe("Timeout", func() {
 			err := rm.Create(ctx, timeoutsAllToAll, store.CreateBy(model.MetaToResourceKey(timeoutsAllToAll.GetMeta())))
 			Expect(err).ToNot(HaveOccurred())
 
-			timeoutMap, err := GetTimeouts(ctx, dataplane, rm)
+			timeoutMap, err := topology.GetTimeouts(ctx, dataplane, rm)
 			Expect(err).ToNot(HaveOccurred())
 			Expect(timeoutMap).To(HaveLen(4))
